requirements.txt
crewai==0.14.4
crewai-tools
langchain==0.1.14
langchain-openai==0.1.1
python-dotenv
uvicorn[standard]
fastapi==0.110.0
pypdf2
docx
tiktoken
requests
google-search-results